LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/body-onerror-compile-error-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ')'

PASS <body onerror> - compile error in <script>
PASS <body onerror> - compile error in <script> (column)
